EVM: Accept Ethereum transactions as native off-ledger requests
This is another step into making EVM a first-class citizen.
Key highlights:

* The wasp node now provides the jsonrpc service at
  `/chain/<chainid>/evm/jsonrpc` (see `packages/webapi/evm`)
* The Ethereum tx is wrapped into an off-ledger request
* The off-ledger request now has an `OffLedgerSignatureScheme` which can
  be one of two types (see `packages/iscp/request*.go`):
    * `isc`: signature is calculated just as before
    * `evm`: signature is provided by the Ethereum tx
* To facilitate estimating gas, there is a new `SimulateCall` function
  that executes a request without committing any changes to the chain
  (see `packages/chain/chainutil/simulatecall.go`).
* Removed the block time feature (evm contracts will be able to send
  timelocked requests anyway).

TODO:

* How to specify the allowance from metamask?
* Gas estimation needs some more work:
    * The proper way to estimate gas is to use binary search modifying
      the gas limit, but for that the call must be executed many
      times, reverting changes between calls (e.g. a call to sandbox.send())
    * Bookkeeping gas (i.e. for maintaining the "ethereum blockchain")
      is not properly calculated. Is it fair for the user to pay for
      that?
* Disabled tests involving NFTs -- how to send an NFT to an ethereum
  address on L2?
* EVM contracts owning and manipulating tokens is untested.
* Update docs

+++ tools/evm/evmcli/http.go
@@ -39,7 +39,7 @@ func (j *JSONRPCServer) getUnlockedAccount() []*ecdsa.PrivateKey {
 	return []*ecdsa.PrivateKey{account}
 }
 
-func (j *JSONRPCServer) ServeJSONRPC(backend jsonrpc.ChainBackend, chainID int) {
+func (j *JSONRPCServer) ServeJSONRPC(backend jsonrpc.ChainBackend, chainID uint16) {
 	evmChain := jsonrpc.NewEVMChain(backend, chainID)
 
 	accountManager := jsonrpc.NewAccountManager(j.getUnlockedAccount())
